requirements.txt
python-dotenv
langchain
langchain-community
langchain-huggingface
faiss-cpu
transformers
huggingface-hub
